+++ apicast/src/threescale_utils.lua
@@ -1,5 +1,6 @@
 
 local redis = require 'resty.redis'
+local env = require 'resty.env'
 
 local _M = {} -- public interface
 
@@ -88,8 +89,8 @@ function _M.required_params_present(f_req, actual)
 end
 
 function _M.connect_redis(host, port)
-  local h = host or os.getenv('REDIS_HOST') or "127.0.0.1"
-  local p = port or os.getenv('REDIS_PORT') or 6379
+  local h = host or env.get('REDIS_HOST') or "127.0.0.1"
+  local p = port or env.get('REDIS_PORT') or 6379
   local red = redis:new()
   local ok, err = red:connect(h, p)
   if not ok then